+++ src/components/Medical/Medical.js
@@ -14,14 +14,13 @@ export default function Medical(props) {
         <section>
           <h3 className='medicalTitle'>Referral</h3>
           <p>
-            If you bring in a referral from your doctor, we will provide you with a superbill upon request in
-            which you can send to your insurance to get a possible reimbursement 
+            If you bring in a referral from your doctor, we will provide you with a superbill upon request that you can send to your insurance to get a possible reimbursement 
             through your insurance company.
           </p>
           <p>
             You can request a referral from your doctor based on your current complaints/issues.
           </p>
-          <p>If you do not bring in a referral, we can still provide you with a superbill, but you have a much higher chance of receiving a reimbursement through your insurance if they accept superbill and have a current referral from them.</p>
+          <p>If you do not bring in a referral, we can still provide you with a superbill, but you have a much higher chance of receiving a reimbursement through your insurance if they accept a superbill and have a current referral from your doctor.</p>
         </section>
 
         <section>
@@ -32,11 +31,11 @@ export default function Medical(props) {
           </p>
           <p>
             We will not give reimbursement if insurance denies your superbill.
-            Through this process, we can not guarantee that insurance will
+            Through this process, we cannot guarantee that insurance will
             accept the superbill, but it is the closest we can provide to make it possible for
             your insurance to cover the massage session.
           </p>
-          <p>The superbill will be an itemized receipt given to the patient that records out-of-network services done. This is required for submitting to your insurance company for out-of-network providers such as our company. This allows the insurance to reimburse the client directly, not the company.</p>
+          <p>The superbill will be an itemized receipt given to the patient that records out-of-network services done. Insurance companes require this for out-of-network providers such as our company. This allows the insurance to reimburse the client directly, not the company.</p>
         </section>
 
         <section>
